// src/types/response.ts
export interface AddressResponse {
    ISO3166_2_lvl3: string; // Contoh: "ID-JW"
    ISO3166_2_lvl4: string; // Contoh: "ID-JI"
    country: string; // Contoh: "Indonesia"
    country_code: string; // Contoh: "id"
    county: string; // Contoh: "Kediri Regency"
    municipality: string; // Contoh: "Purwoasri"
    postcode: string; // Contoh: "64154"
    region: string; // Contoh: "Java"
    state: string; // Contoh: "East Java"
    village: string; // Contoh: "Karangpakis"
}